Reduce unneeded abstraction.
Before this changeset the graphical repsentation of the chessboard was
wrapped in the ChessboardWidget class. This wasn't needed as it would
mean duplicating the API between ChessboardWidget and the graphical
chessboard class.

The ChessboardWidget class has been removed and its functions has been
transfered to the MainWindow class.

Project file has been modified accordingly.


git-svn-id: https://sloppygui.svn.sourceforge.net/svnroot/sloppygui/trunk@9 1916b7bc-2d4d-0410-bf2b-f14ee8b928b2