Implement initial support for program settings.
This commit introduces initial support for saving and loading program
settings.

Only the chess engine configurations are saved currently.

Program settings are saved to the user's home directory. The specific
path depends on the operating system used. The format of the settings
file however is identical across all platforms.

This commit has been labeled initial because the settings support is far
from perfect. Currently it requires using number of parameters
such as the application's name, organization and the file format. In the
future the settings accessed through a dedidated configuration manager.