Complete the 'gui' and 'lib' source code separation.
The source is now divided into two projects gui and lib. Each of these projects
are independent from the build system's view. The gui project however
requires the lib to be built but each project can be built separately.

The lib project creates a dynamic library called "chess". This chess
library only depends on the QtCore library meaning that it doesn't contain
any GUI code.

The lib project also contains unit tests which require the library to be built.
Because the tests now dynamically link to the chess library, building the
tests is a lot faster.

The gui project creates the old GUI interface but now requires linking
to the "chess" library.