Add PlainTextLog GUI component.
The PlainTextLog's purpose is to act as logging component in the UI. It is
inherited from QPlainTextEdit.

PlainTextLog adds features to clear the log and save its current content to
a file. Both of these features can be accessed from the context menu.

In order to use the save feature, the 'saveLogToFileRequest' signal must
be catched and its slot must implement to do the actual saving.

PlainTextLog is read-only by default.