One configuration class to rule them all
Using two different classes for the same thing is a bad idea. Try to use
EngineConfiguration everywhere.

The remaining EngineSettings usage is for the time control which should be
removed too.

EngineConfiguration::protocol() was changed to int to avoid dependency
cycle between EngineConfiguration and ChessEngine classes. The
representation of chess engine protocols will change in the near future
anyway.