Add a StartImmediately mode to GameManager
GameManager::newGame() now has to modes for starting games:

* StartImmediately, which ignores the queue and starts the game
  right away. This should be used in the GUI, for games started
  manually by the user.
* Enqueue, which adds the game to the queue and starts it when
  a game slot becomes free. This mode is suitable for batched
  engine-engine matches and tournaments.

The two modes can be used at the same time, ie. queued and non-queued
games can be played simultaneously. This is a good feature if the user
wants to play a game of chess while an engine-engine tournament is
running in the background.