Infinite time control
Infinite time control is a good choice if the user wants to give himself
infinite thinking time. It's usually not so suitable for engines because
they'll just think forever until the user forces them to move. However,
infinite time control is still supported even in cutechess-cli because
it can be used in combination with depth and node limits.

In infinite time control the GUI's chess clocks start at 0 and count the
total time used.