Major chessboard overhaul
This monster commit replaces the old model-view graphical board with
a Graphics View implementation. To make it all work, some changes and
new features in the internal board classes were also needed.

To summarize this commit:
* BoardScene is now THE graphical board. BoardView displays the scene.
* Chess::Board and its subclasses are now normal C++ classes instead of QObjects.
* The board transitions are retrieved from the internal board by passing a
  Chess::BoardTransition object to Chess::Board::makeMove().
* Chess::Board has a copy() method which returns a copy of the board. This was
  done mainly because BoardScene needs its own copy of the internal board.
* There's no longer any need for a separate promotion dialog because
  there's a PieceChooser class that takes care of that. In the future
  PieceChooser can also be used for editing board positions.
* Moves are animated, but any new transition in the scene will immediately
  stop the animations. That is, an animation does not slow down the game if
  moves are made faster than the animation take to complete.

In the future the new classes may need some cleanup, and some of the lost
features like highlighting legal moves need to be restored. But this is a
very functional base.