Add EngineOptionDelegate class
EngineOptionDelegate class creates editors such as comboboxes and
spinboxes for engine options in Qt item views. EngineOptionDelegate
works by trying to convert data from the model to known engine option
classes. If the conversion succeeds, an editor is created based on that
engine option. If the conversion fails the data is passed on to
QStyledItemDelegate for standard editor creation.