Making the first move starts the game
In games where a human player is first to move, the game is started when
the first move is made. This required the following changes:

* ChessGame has a pause() slot which pauses the game after any pending
moves have been made, and a resume() slot which resumes the game.
* When the user makes the first move, HumanPlayer tells the ChessGame
object to resume by emitting the wokeUp() signal. When the game resumes
the HumanPlayer object plays the move.
* Games can be paused before they're even started