Add TileLayout class
TileLayout is a layout manager based on the FlowLayout example.
It tries to fill the available space with its items as efficiently
as possible by tiling them in columns and rows. All items in the
layout will have the same size, and the aspect ratio of their
sizeHint shouldn't differ too much.

The tiling algorithm is far from perfect, but good enough to give
optimal results in most cases.