Fixed a serious bug in Cutechess-cli
The command line options were stored in a QMultiMap which sorts the
items based on the option names. This means that for the new "-engine"
options there was no guarantee that the first "-engine" would actually
end up as the first engine in the tournament, which is a very bad thing
for those who optimize their engine with CLOP.

This commit replaces the multimap with a list, guaranteeing that options
are processed in insertion order.