Use index-based filtering in PgnGameEntryModel
Since there's no index-based version of QtConcurrent::filtered(), we're
using an extra QVector for simply storing a list of indexes that can be
passed to QtConcurrent::filtered(). This allows us to easily return the
original index based on the filtered item's row. And turns out this
doesn't use more memory than the previous implementation and is a bit
faster too.