Fix GameDatabaseDialog's game indexes
Now databaseIndexFromGame() returns the correct index because we're
using the original game index instead of the filtered one. And
m_selectedDatabases is now a QMap to make sure that the database indexes
are always retrieved in the right order. This increases performance a
bit too, because we don't have to call qSort() any more.