Implement the base for new move list
This commit implements the base for the new move list. The new move list
is based on QTextEdit that is highlighted using a custom syntax
highlighter. The highlighter is also responsible for tokenizing the list
of moves so that the information can be used later.

Clicking a move or a comment does not do anything currently. The signaling
method (with correct information) for these events is already in place
however.