Use Q_DECLARE_TR_FUNCTIONS and avoid global objects
Calling the static QObject::tr() from non-QObject classes isn't good
because it results in the wrong translation context.

As I was trying to fix the translation context of the header names in
model classes I ended up removing the static QStringList objects
completely and using a switch statement in headerData() instead (as we
already did in data()).