Keep tournament's PGN output file open
Opening and closing the PGN output file for each game was causing
problems on Windows with fast time controls. So let's just open the file
once and stream the PGN output there. This is also a lot more efficient.
The file is opened (created) when the first game is ready to be written
to disk.