Avoid using QDir::setCurrent()
Calling QDir::setCurrent() when building a chess engine required us to
protect the code block with a mutex lock. But even that wasn't enough
because at the same time CuteChess could read or write to some file (eg.
a PGN collection) with a relative path.

Now the problem is solved cleanly in the EngineProcess class.