Flush PGN output buffer when saving a game
The fix was suggested by Peter Österlund:
"I have encountered a small problem related to how games are saved. The PGN
output file is now kept open, and it uses internal buffering. This means that
if you try to read the file during a match, the most recently finished game
is typically truncated."