Begin the long road to Qt5 only
Qt4 end-of-life is approaching so it's time to start migrating to Qt5
only. Debian stable has Qt 5.3 so use it as the lowest required version.

We can start using Qt5 features now up until 5.3. Also, C++11 is now
supported but which limited set of C++11 should be used should be
discussed separately.