Add configuration file for Travis CI
In the near future Travis CI will be used to test building Cute Chess
with Qt 5.3, 5.4 and 5.5 after each commit. The aim is to prevent a
situation where Cute Chess would not built on any of the supported
Qt versions.

Later the testing will be expanded to cover all unit tests in Cute
Chess.