Make call to std::pow unambigious
On OS X 10.11 (El Capitan) with LLVM / Clang the build fails due to
std::pow arguments being ambigious. This is because the std::pow
function is defined by using templates in LLVM's standard C++ library. This
isn't the case on all operating systems which is the reason why our CI
system didn't catch this.

Cast the first argument explicitly to double to make the function call
unambigious.

The CI system will be expanded to cover OS X in the near future to catch
situations like this.