Implement runtime consistency checks for models
This commit adds support for checking the consistency of
QAbstractItemModel based model classes. The actual consistency checks
are implemented in ModelTest class which is part of Qt:
https://wiki.qt.io/Model_Test

The ModelTest class had to be modified to compile with Qt5. Some
excessive debug output was disabled as well but other than that
modifications have been kept as small as possible.

Checks have been added to more complex model classes. If a check fails a
message is printed to the standard output. Checks are enabled by default
even in release builds. This is something that will be revised later.