gui: start moving to non-blocking dialogs
The problem of using QFileDialog's static functions such as
getOpenFileName is that they block UI events in all dialogs. This can be
easily verified by opening the Game Database window and opening a file
in the game window. The result is that the Game Database window cannot
be interacted with until the file dialog is closed. The problem is not
limited to file dialogs but all dialogs that use the exec() method.

The aim is to have have majority of the dialogs non-blocking or 'window
modal' so that other windows can be interacted with at all times. This
requires that dialogs are created with open() method. A little more code
is required to handle accepting the dialog because it will happen
asynchronously. C++11 lambdas help with this.

The first dialog to use this async model is the log view's save log
dialog. This commit moves handling of saving the file to PlainTextLog
class and uses the QDialog::open() method to create a 'window modal'
dialog. This will only block the current game window and other windows
such as Game Database can be interacted with.