Add Worker and WorkerPool classes
The Worker and WorkerPool classes are used to unify the running of long
background tasks. The classes allow a clear separation of implementation
and running the background task.

The plan is to remove the ThreadedTask class in the future.

Modify the PGN importer to use the new Worker classes. The rest of the
classes currently using ThreadedTask will follow.