Replace WorkerPool with QThreadPool
For the use cases we have now QThreadPool works fine. If there's any
need for more specialized functionality in the future we can replace
it with something else.

The Worker class remains essentially the same except it's now inherited
both from QObject and QRunnable. The QRunnable::run() method now
implements the old Worker::start() keeping the functionality the same.