Add build.sr.ht CI manifests
This commit adds support for Sourcehut (https://sourcehut.org) CI,
https://build.sr.ht. The following operating systems are supported:

* Debian stable
* Arch Linux
* FreeBSD 12.x

This gives us a pretty good coverage of different Qt versions and OSes.

Sourcehut CI is not used by default -- only if the code is pushed to the
relevant Sourcehut git repositories.

While here, remove the `-Wshadow` compiler flag as GCC's output clobbers
up the build logs. Clang is fine in this regard, but now majority of our
builds are compiled with GCC.