Add LGTM queries for spotting potential problems
This commit adds two LGTM queries for spotting potential problems and
areas of improvement in the source code. These queries are automatically
enabled in the Github pull request phase but can be also run manually
through the LGTM query console.

The two added queries try to spot the following:

1) Use of QFileInfo.exists("filename") instead of
   QFileInfo::exists("filename"). QFileInfo:exists is faster.

2) Returning a void expression from a function

Both of these queries return results in the current master branch.

The checks are based on those by Clazy but written from
scratch. The plan is to have more checks in the future.