Read halfmove clock and fullmove number from epd opening files
A position in epd format doesn't store halfmove clock and fullmove number in the fen part, but they can still be specified (optionally) through the hmvc and fmvn opcodes. Use this when converting epd to fen.

As a side effect, if hmvc and fmvn are not set, we now return a fen with halfmove clock and fullmove number set to 0 and 1, since returning a fen without those is not strictly correct.