add support for GitHub Actions
This commit adds support for GitHub Actions:
https://help.github.com/en/actions

GitHub Actions allow automated workflows to be executed when, for
example, pull requests are created.

This commit uses GitHub Actions to build and test the Cute Chess code
base using Windows, Ubuntu Linux and MacOS. The following Qt versions
are used with all operating systems: 5.11, 5.12, 5.13 and 5.14. The
build and test cycle is run for all pull requests and pushes.

The plan is to eventually phase out AppVeyor and Travis CI.