add initial clang-format configuration
This commit adds an initial version of the clang-format configuration.
Clang-format is a tool that implements automatic source code formatting
based on the clang compiler.

The formatting rules follow the Qt conventions:

https://github.com/qt/qt5/blob/dev/_clang-format

However, there are some differences such as the pointer asterisk being
attached to the type rather than the variable in our style.

Adhering to the Qt style results in many changes in the code formatting
making the style more "compact". It's up to debate whether the style in
this commit is the style chosen for the future. The important thing is
that we have a style that can be applied with a tool.

The current code base will not be formatted with these rules as it would
result in large, unnecessary diffs.