refactor xboard feature and pv parsing
Extract the feature and principal variation parsing in to their own
methods. This will make it easier to create unit tests for these
implementations.

The current API is inelegant mainly due to mechanical copy-paste of the
implementations to parsePv() and parseFeatures() methods, respectively.
The API will be improved (and the parsing rewritten) once unit tests
are in place.