github ci: fix qt version conditional
The 'if' conditional cannot be used the way we are currently using it.
The current implementation results in the following warning:

Unexpected input(s) 'if', valid inputs are ['dir', 'version', ...

The conditional must be at the top level and not under 'with'. This
means that the Qt installation step must be split into two: Qt 5.x and
Qt 6.x.