Introduce sub-rounds and sub-sub-rounds to tournament round counting and PGN output.
Resolves #249

Use subround counting for round-robin tournaments. This fixes the round
repetition policy where the PGN round does not correspond to a round
specified by the -rounds option. Now the PGN round tag is constant for
the whole cycle of sub-rounds.

Resolves #641

To accomplish a more fine-grained control of openings two new policy
options are introduced to OpeningPolicy:

 1. SubRoundPolicy which shifts openings for a new sub-round.
 2. KeepPolicy which always keeps the same opening.

Thus there a policies to shift or not shift openings for the levels
of a tournament (game, encounter, sub-round, round, tournament).

For pyramid tournaments the function of the round counter is now
distributed between round and sub-round counter.

In knockout tournaments the sub-round is constant for a pair,
the sub-sub-round counts the games of a pair.

In gauntlet tournaments the sub-round counter tracks the gauntlet
(seeded) players, the sub-sub-round counter individual games.

The "round" field was changed from int to QString in PgnGame.