add cmake support
This commit adds support for building Cute Chess with CMake.

QMake, which is the current build system, will be deprecated in the
future so at some point we need to make the transition to CMake, which
is practically the only choice forward. CMake is used to build Qt 6.

In its current form QMake and CMake exist in parallel: the project can
be build in either two systems but the plan is to remove QMake support
completely.

The default CI build will be with CMake.

Changes to the source tree were intended to be kept as minimal as
possible to accommodate CMake:

- Modeltest was disabled from the build. Debug builds on CMake need more research.

- Tests data was moved to its own dedicated directory, lib/tests/data,
  in order to make declaring unit tests easier on CMake.

- Ui files were modified to search their header files in a consistent
  location on both QMake and CMake.

The default build type is release with debug info.

The lib project is always built statically -- there's no plans to add
dynamic linking due to lack resources to keep up with API / ABI
stability within the library.