add release-tool
The purpose of the `release-tool` is to help ensure consistent Cute
Chess release. The script will check common issues during release
generation and abort if errors are observed.

When a new release is ready to be shipped, the `release-tool` will be
run with the new version, for example: `release-tool 1.3.0`.

Inspired by the `release-tool` in KeepassXC repository and
https://drewdevault.com/2019/10/12/how-to-fuck-up-releases.html

Fixes #602