re-enable local CodeQL queries (#798)
These local queries were developed some years ago but were disabled when
the LGTM service (predecessor to CodeQL) was deprecated.

Reference the directory where the local queries are located
with the `queries` parameter in the `codeql/init` step. Add a QL pack
configuration so that CodeQL understand this directory contains the
queries.